feat: plugin-clean - cleanup / wait (#6)
* feat: cleanup and wait

* cleanup on generateBundle

* introduces internal deleted flag

* fix deleteOnce test

* convert to multi config plugin, add wait/check subdir logic

* remove temp file

+++ plugin-clean/src/index.ts
@@ -1,12 +1,15 @@
 import fs from 'fs/promises';
 import path from 'path';
-import { PluginHooks, OutputOptions, RollupOptions } from 'rollup';
+import { NormalizedInputOptions, NormalizedOutputOptions, PluginHooks, RollupOptions } from 'rollup';
 import { CleanPluginOptions } from './types';
 import { createLogger, LogLevel } from '@niceties/logger';
 import { getOptions } from '@rollup-extras/utils/options';
+import { multiConfigPluginBase } from '@rollup-extras/utils/mutli-config-plugin-base';
 
 export default function(options: CleanPluginOptions = {}) {
-    const deleted = new Map<string, Promise<void>>();
+    const inProgress = new Map<string, Promise<void>>();
+    const hasChildrenInProgress = new Map<string, Promise<void>>();
+    let deleted = false;
 
     const normalizedOptions = getOptions(options, {
         pluginName: '@rollup-extras/plugin-clean',
@@ -17,18 +20,38 @@ export default function(options: CleanPluginOptions = {}) {
     const { pluginName, deleteOnce, verbose, outputPlugin } = normalizedOptions;
     let { targets } = normalizedOptions;
 
-    const pluginInstance = {
-        name: pluginName
-    } as Partial<PluginHooks>;
+    const instance = multiConfigPluginBase(false, pluginName, cleanup);
+    const baseAddInstance = (instance as Required<typeof instance>).api.addInstance;
+    const baseRenderStart = (instance as Required<typeof instance>).renderStart;
 
     if (outputPlugin) {
-        pluginInstance.renderStart = renderStart;
+        instance.renderStart = async function (options: NormalizedOutputOptions, inputOptions: NormalizedInputOptions) {
+            baseRenderStart.call(this, options, inputOptions);
+            await renderStart(options);
+        };
     } else {
-        pluginInstance.buildStart = buildStart;
-        pluginInstance.options = optionsHook;
+        instance.buildStart = buildStart;
+        instance.options = optionsHook;
     }
 
-    return pluginInstance;
+    instance.api.addInstance = () => {
+        const instance = baseAddInstance() as never as PluginHooks;
+        const baseRenderStart = (instance as Required<typeof instance>).renderStart;
+
+        if (outputPlugin) {
+            instance.renderStart = async function (options: NormalizedOutputOptions, inputOptions: NormalizedInputOptions) {
+                baseRenderStart.call(this, options, inputOptions);
+                await renderStart(options);
+            };
+        } else {
+            instance.buildStart = buildStart;
+            instance.options = optionsHook;
+        }
+    
+        return instance;
+    };
+
+    return instance;
 
     async function optionsHook(config: RollupOptions) {
         if (!targets && config) {
@@ -39,12 +62,18 @@ export default function(options: CleanPluginOptions = {}) {
     }
 
     async function buildStart() {
+        if (deleted) {
+            return;
+        }
         if (targets) {
             await Promise.all(targets.map(removeDir));
         }
     }
 
-    async function renderStart(options: OutputOptions) {
+    async function renderStart(options: NormalizedOutputOptions) {
+        if (deleted) {
+            return;
+        }
         if (!targets) {
             if (options.dir) {
                 await removeDir(options.dir);
@@ -56,11 +85,27 @@ export default function(options: CleanPluginOptions = {}) {
 
     async function removeDir(dir: string) {
         const normalizedDir = normalizeSlash(path.normalize(dir));
-        if (deleteOnce && deleted.has(normalizedDir)) {
-            return outputPlugin && deleted.get(normalizedDir);
+        if (inProgress.has(normalizedDir)) {
+            return outputPlugin && inProgress.get(normalizedDir);
         }
-        const removePromise = doRemove(normalizedDir);
-        deleted.set(normalizedDir, removePromise);
+        let removePromise: Promise<void>;
+        let parentsInProgress: string[];
+        if (hasChildrenInProgress.has(dir)) {
+            removePromise = Promise.resolve(hasChildrenInProgress.get(dir))
+                .then(() => doRemove(normalizedDir));
+        } else if ((parentsInProgress = Array.from(parentDirs(dir)).filter(item => inProgress.has(item))).length > 0) {
+            return inProgress.get(parentsInProgress[0]);
+        } else {
+            removePromise = doRemove(normalizedDir);
+        }
+        inProgress.set(normalizedDir, removePromise);
+        for(const parentDir of parentDirs(dir)) {
+            if (!hasChildrenInProgress.has(parentDir)) {
+                hasChildrenInProgress.set(parentDir, removePromise);
+            } else {
+                hasChildrenInProgress.set(parentDir, Promise.all([removePromise, hasChildrenInProgress.get(parentDir)]) as never as Promise<void>);
+            }
+        }
         return removePromise;
     }
 
@@ -75,6 +120,12 @@ export default function(options: CleanPluginOptions = {}) {
             logger.finish(`failed cleaning '${normalizedDir}'`, loglevel, e);
         }
     }
+
+    function cleanup() {
+        inProgress.clear();
+        hasChildrenInProgress.clear();
+        deleted = deleteOnce;
+    }
 }
 
 function normalizeSlash(dir: string): string {
@@ -82,4 +133,10 @@ function normalizeSlash(dir: string): string {
         return `${dir.substring(0, dir.length - 1)}`;
     }
     return dir;
+}
+
+function *parentDirs(dir: string) {
+    while(((dir = path.dirname(dir)) !== '.') && (dir !== '/')) {
+        yield dir;
+    }
 }
